+++ buildcontext/local.go
@@ -91,7 +91,14 @@ func (lr *localResolver) resolveLocal(ctx context.Context, gwClient gwclient.Cli
 	if gwClient != nil {
 		if _, isTarget := ref.(domain.Target); isTarget {
 			noImplicitIgnore := bf.ftrs != nil && bf.ftrs.NoImplicitIgnore
-			excludes, err := readExcludes(ref.GetLocalPath(), noImplicitIgnore)
+
+			useDockerIgnore := isDockerfile
+			ftrs := features.FromContext(ctx)
+			if ftrs != nil {
+				useDockerIgnore = useDockerIgnore && ftrs.UseDockerIgnore
+			}
+
+			excludes, err := readExcludes(ref.GetLocalPath(), noImplicitIgnore, useDockerIgnore)
 			if err != nil {
 				return nil, err
 			}